Renamed file, removed old query sample
