+++ src/components/boxes.tsx
@@ -4,8 +4,8 @@ interface BoxContent {
 
 export function FormBox({children}: BoxContent) {
     return(
-        <div className="grid gap-4 flex-col items-center bg-white p-5 rounded-3xl ">
-            {children}
+        <div className="grid gap-4 flex-col items-center border-gray-600 bg-white p-5 rounded-3xl ">
+        {children}
         </div> 
     )
 }
